pyproject.toml
[tool.poetry]
name = "fastapi-admin-next"
version = "0.1.0"
description = ""
authors = ["MahmudulHassan5809 <mahmudul.hassan240@gmail.com>"]
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic = "^2.10.3"
sqlalchemy = "^2.0.36"
jinja2 = "^3.1.4"
itsdangerous = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pre-commit = "^4.0.1"
black = "^24.10.0"
pylint = "^3.3.2"
flake8 = "^7.1.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
scriv = {extras = ["toml"], version = "^1.5.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
